data_analyse-master/requirements.txt
streamlit
streamlit_chat
streamlit-feedback
langchain
langchain-community
langchain-openai
unstructured
python-dotenv
elasticsearch
langchain_experimental
matplotlib
seaborn
scipy
wikipedia
scikit-learn
sentence_transformers
unstructured_inference
unstructured_pytesseract
libmagic
tesseract
pinecone-client
pillow_heif
numpy==1.23.5
langchain-pinecone
chromadb
pikepdf
pypdf